Device and SwapChain rough implemention Pipeline still missing so running it wont do lots of holes in the design for application interaction

+++ Parabol/StringUtils.cpp
@@ -56,6 +56,6 @@ void NonCpyStringContainer::add(const std::vector<std::string>& rref) {
 		this->add(ref);
 	}
 }
-uint32 NonCpyStringContainer::size() {
+size_t NonCpyStringContainer::size() {
 	return this->strs.size();
 }
